test(dtls): add a basic DTLS test, refactor existing test
